Add constants index
