--- app/components/views/GetStartedPage/CreateWallet/CreateWalletForm/ExistingSeed/constants.js
@@ -1,4 +1,0 @@
-export const WORDS = "words";
-export const HEX = "hex";
-export const POSITION_ERROR = "not valid at position";
-export const MISMATCH_ERROR = "checksum mismatch";
