+++ app/wallet/index.js
@@ -3,7 +3,6 @@ import { TransactionDetails } from "../middleware/walletrpc/api_pb";
 export * from "./app";
 export * from "./control";
 export * from "./client";
-export * from "./constants";
 export * from "./daemon";
 export * from "./loader";
 export * from "./message";
